requirements.txt
flask
openai==0.28.1
requests
python-dotenv
unittest2==1.1.0
pytest==7.4.3
pytest-pythonpath==0.7.4